+++ src/Helpers/helpers.php
@@ -1,0 +1,6 @@
+<?php
+
+function projects_config_published()
+{
+    return file_exists(config_path('projects.php'));
+}