+++ CHANGELOG.md
@@ -1,3 +1,24 @@
 # Changelog
 
 All notable changes to `laravel-projects` will be documented in this file
+
+## 0.0.1 - 29-03-2020
+
+- Initial release
+
+## 0.0.2 - 05-04-2020
+
+- Setup Status and Priority Models
+- Tasks have Statuses and Priorities
+- Projects have Statuses
+
+## 0.0.3 - 06-04-2020
+
+- Setup README.md
+- Publishing config and seeder
+- Statuses and Priorities scopes
+
+### TODO/WIP:
+
+- Scopes
+- Pricing sum values